# Conflicts:
#	pandas/core/series.py

+++ pandas/core/tools/datetimes.py
@@ -307,7 +307,7 @@ def _convert_and_box_cache(
     """
     from pandas import Series
 
-    result = Series(arg).map(cache_array)
+    result = Series(arg, dtype=cache_array.index.dtype).map(cache_array)
     return _box_as_indexlike(result._values, utc=False, name=name)
